pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diagrambot"
version = "0.1.0"
description = """
Create mermaid and graphviz diagrams using Shinyrealtime and the OpenAI Realtime 
API. This package provides a Shiny for Python interface for generating various 
types of diagrams through voice commands.
"""
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Sam Parmar"}
]
keywords = ["shiny", "python", "diagrams", "visualization", "interactive"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "shiny>=0.6.0",
    "plotly>=5.0.0",
    "networkx>=2.8.0",
    "pandas>=1.3.0",
    "matplotlib",
    "plotnine",
    "seaborn",
    "shinychat",
    "python-dotenv",
    "shinyrealtime",
]

[project.urls]
Homepage = "https://github.com/parmsam/diagrambot-py"
Repository = "https://github.com/parmsam/diagrambot-py"